airflow/requirements.txt
astronomer-cosmos[dbt-redshift]==1.9.0
apache-airflow-providers-amazon==9.4.0
